Refactor tests, split async builder and client
