+++ engineio/src/asynchronous/client/mod.rs
@@ -1,2 +1,4 @@
 mod async_client;
-pub use async_client::{Client, ClientBuilder};
+mod builder;
+pub use async_client::Client;
+pub use builder::ClientBuilder;
